requirements/base.txt
--constraint=constraints.txt

Jinja2
jmespath
msgpack>=1.0.0
PyYAML
MarkupSafe
requests>=2.31.0 ; python_version < '3.8'
requests>=2.32.0 ; python_version >= '3.8'
distro>=1.0.1
psutil>=5.0.0
packaging>=21.3
looseversion
tornado>=6.3.3
aiohttp>=3.9.0
croniter>=0.3.0,!=0.3.22; sys_platform != 'win32'

# We need contextvars for salt-ssh.
# Even on python versions which ships with contextvars in the standard library!
contextvars

setproctitle>=1.2.3
timelib>=0.2.5
pyopenssl>=24.0.0
python-dateutil>=2.8.1
python-gnupg>=0.4.7
cherrypy>=18.6.1
importlib-metadata>=3.3.0
cryptography>=42.0.0

# From old requirements/static/pkg/linux.in
rpm-vercmp; sys_platform == 'linux'

# From old windows.txt requirements file
gitpython>=3.1.37; sys_platform == 'win32'
lxml>=4.6.3; sys_platform == 'win32'
pymssql>=2.2.1; sys_platform == 'win32'
pymysql>=1.0.2; sys_platform == 'win32'
pythonnet>=3.0.1; sys_platform == 'win32'
pywin32>=305; sys_platform == 'win32'
wmi>=1.5.1; sys_platform == 'win32'
xmltodict>=0.13.0; sys_platform == 'win32'
